[step5] refactor domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현 refactor 클래스명 변경
