+++ src/main/java/racing/view/OutputView.java
@@ -1,9 +1,5 @@
 package racing.view;
 
-import java.util.List;
-import racing.race.Car;
-import racing.race.Cars;
-
 public class OutputView {
 
   public static void prinln() {
@@ -30,12 +26,7 @@ public class OutputView {
     System.out.println();
   }
 
-  public static void winner(List<Car> cars) {
-    StringBuilder winnerNames = new StringBuilder();
-    for (Car car : cars) {
-      winnerNames.append(car.name()).append(", ");
-    }
-    System.out.println(winnerNames.delete(winnerNames.length() - 2, winnerNames.length())
-        + "가 최종 우승했습니다.");
+  public static void winner(String names) {
+    System.out.println(names + "가 최종 우승했습니다.");
   }
 }
